Moved to react-services, changed name, traslates comments
