+++ public/react-services/logger-service.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { expect } from 'chai';
-import { loggerService } from './logger-service';
+import { loggerService } from './logger-services';
 
 describe('Spec LoggerService', function () {
   it('it exists', () => {
